# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Documentation",
    "description": "\n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWTAuth": {
        "type": "http",
        "description": "JWT Authentication Scheme",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "GeoLocation": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "latitude": {
            "description": "Latitude must be between -90 and 90",
            "type": "number"
          },
          "longitude": {
            "description": "Longitude must be between -180 and 180",
            "type": "number"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ],
        "additionalProperties": false,
        "id": "GeoLocation"
      },
      "AnyObject": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "__": {
            "type": "string"
          }
        },
        "required": [
          "__"
        ],
        "additionalProperties": false,
        "id": "AnyObject"
      },
      "PaginationMetaData": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ],
        "additionalProperties": false,
        "id": "PaginationMetaData"
      },
      "BusData": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "licensePlate": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "metadata": {
            "$ref": "#/components/schemas/AnyObject"
          }
        },
        "required": [
          "id",
          "licensePlate",
          "capacity",
          "metadata"
        ],
        "additionalProperties": false,
        "id": "BusData"
      },
      "RouteData": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "name": {
            "type": "string"
          },
          "startLocation": {
            "$ref": "#/components/schemas/GeoLocation"
          },
          "endLocation": {
            "$ref": "#/components/schemas/GeoLocation"
          },
          "metadata": {}
        },
        "required": [
          "id",
          "name",
          "startLocation",
          "endLocation",
          "metadata"
        ],
        "additionalProperties": false,
        "id": "RouteData"
      },
      "TimeTable": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string"
          },
          "departureTime": {
            "type": "string"
          }
        },
        "required": [
          "dayOfWeek",
          "departureTime"
        ],
        "additionalProperties": false,
        "id": "TimeTable"
      },
...
    }
  },
  "paths": {
...
    "/api/schedules/": {
      "get": {
        "tags": [
          "SchedulesController"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "description": "Search term to filter results",
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "default": 1,
              "description": "Page number, minimum is 1",
              "type": "number",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "default": 10,
              "description": "Number of items per page, minimum is 1 and maximum is 100",
              "type": "number",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "data": {
                      "$schema": "https://json-schema.org/draft/2020-12/schema",
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid",
                                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                              },
                              "bus": {
                                "$ref": "#/components/schemas/BusData"
                              },
                              "times": {
                                "type": "array",
                                "items": {
                                  "$ref": "#/components/schemas/TimeTable"
                                }
                              },
                              "route": {
                                "$ref": "#/components/schemas/RouteData"
                              },
                              "meta": {},
                              "startDate": {
                                "type": "string",
                                "format": "date-time",
                                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                              },
                              "endDate": {
                                "type": "string",
                                "format": "date-time",
                                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                              },
                              "startTime": {
                                "type": "string",
                                "format": "time",
                                "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?$"
                              },
                              "type": {
                                "default": "MORNING",
                                "type": "string",
                                "enum": [
                                  "MORNING",
                                  "AFTERNOON"
                                ]
                              }
                            },
                            "required": [
                              "id",
                              "bus",
                              "times",
                              "route",
                              "meta",
                              "startDate",
                              "endDate",
                              "startTime",
                              "type"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaData"
                        }
                      },
                      "required": [
                        "data",
                        "meta"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "SchedulesController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "busId": {
                    "type": "string"
                  },
                  "routeId": {
                    "type": "string"
                  },
                  "driverId": {
                    "type": "string"
                  },
                  "times": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/TimeTable"
                    }
                  },
                  "meta": {},
                  "startDate": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                  },
                  "startTime": {
                    "type": "string",
                    "format": "time",
                    "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?$"
                  },
                  "type": {
                    "default": "MORNING",
                    "type": "string",
                    "enum": [
                      "MORNING",
                      "AFTERNOON"
                    ]
                  }
                },
                "required": [
                  "busId",
                  "routeId",
                  "driverId",
                  "times",
                  "startDate",
                  "endDate",
                  "startTime",
                  "type"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "data": {
                      "$schema": "https://json-schema.org/draft/2020-12/schema",
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                        },
                        "bus": {
                          "$ref": "#/components/schemas/BusData"
                        },
                        "times": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TimeTable"
                          }
                        },
                        "route": {
                          "$ref": "#/components/schemas/RouteData"
                        },
                        "meta": {},
                        "startDate": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        "endDate": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        "startTime": {
                          "type": "string",
                          "format": "time",
                          "pattern": "^(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?$"
                        },
                        "type": {
                          "default": "MORNING",
                          "type": "string",
                          "enum": [
                            "MORNING",
                            "AFTERNOON"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "bus",
                        "times",
                        "route",
                        "meta",
                        "startDate",
                        "endDate",
                        "startTime",
                        "type"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
...
  }
}